Views: App: Add 'Interface' item  on configurantion menu

Views: Configuiration-interface: Add menu item for interface configurations


Views: Configuiration-general: Remove interface configurations
